requirements.txt
fastapi
uvicorn
pydantic
redis
pytest
langfuse
python-dotenv
openai